+++ client/web-admin/src/components/Dashboard/ListAssignedJurisdictions/index.js
